lib/models/Heatmap_Service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.0.0
sentinelhub>=3.9.0
python-dotenv>=1.0.0
